third_party/blink/web_tests/platform/linux-chrome/external/wpt/battery-status/page-visibility.https-expected.txt

This is a testharness.js-based test.
[FAIL] battery status events only fire on visible pages
  assert_equals: expected 0.6 but got 0.55
Harness: the test ran to completion.
